requirements.txt
ariadne==0.13.0
Flask==1.1.2
Flask_Bcrypt==0.7.1
Flask_Mail==0.9.1
Flask_SQLAlchemy==2.5.1
PyJWT==2.3.0
requests==2.25.1
cryptography==35.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.2
pytz==2021.3
